inference/pyproject.toml
[project]
name = "inference"
version = "0.1.0"
description = "Inference for SJSU parking predictions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.123.0",
    "joblib>=1.5.2",
    "jupyter>=1.1.1",
    "numpy>=2.0.2",
    "pandas>=2.3.3",
    "psycopg[binary,pool]>=3.3.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.6",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.45",
]
